sr(api): move create-training in devcomp
